requirements.txt
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
seaborn==0.12.2
pycaret==3.1.0
missingno==0.5.2
plotly==5.17.0
matplotlib==3.7.2
boto3==1.28.37
python-dotenv==0.21.1
